Implement the 'as32Bit[Un]signedValue' primitives.
Java SOM crashes on big integers, so I have defined what I believe to be
appropriate semantics for them which appears to be not to perform sign
extension in the conversion. For example:

  ((1 << 60) - 1) as32BitUnsignedValue

in Java SOM returns 4294967295, which is bigger than a signed 32 bit can
represent. This commit preserves that behaviour.