Implement the 'Shr' primitive.
Java SOM seems to define "shift right" as "treat the int as unsigned, shift
right, then treat as signed". This is fairly easy for us in the case of
word-sized integers, but awkward to implement for negative big ints -- I've left
that for the future (if anyone should ever encounter such a case, which might be
quite some time away!).