Implement 'perform:[withArguments:]inSuperclass:'.
These turn out to be quite easy now that we've factored out the "perform"
function. Note a) that we don't check that the class specified is a superclass
of the current class (I couldn't see a correctness advantage in doing so) b)
Java SOM doesn't implement perform:withArguments:inSuperclass, but I felt it was
obvious enough what it should do that I did it anyway.