Use SmartString.
The SmartString library avoids allocating memory for strings if they're small
enough to fit inside the existing String struct. This is quite appealing for an
interpreter where in general there are quite a lot of small strings that are
critical for performance (e.g. function names; though, admittedly, Smalltalk has
many unusually long function names).

This commit is a fairly simple port (note that we generally import `SmartString`
as `String` i.e. `String` no longer means `std::string::String` as per normal).
Here's how long the test suite takes on my desktop before this commit:

  $ multitime -n 30 -s0 -q target/release/yksom --cp SOM/Smalltalk/ SOM/TestSuite/TestHarness.som
  ===> multitime results
  1: -q target/release/yksom --cp SOM/Smalltalk/ SOM/TestSuite/TestHarness.som
              Mean        Std.Dev.    Min         Median      Max
  real        0.214       0.002       0.210       0.214       0.220
  user        0.318       0.028       0.260       0.310       0.360
  sys         0.015       0.011       0.000       0.015       0.040

And here's after:

  $ multitime -n 30 -s0 -q target/release/yksom --cp SOM/Smalltalk/ SOM/TestSuite/TestHarness.som
  ===> multitime results
  1: -q target/release/yksom --cp SOM/Smalltalk/ SOM/TestSuite/TestHarness.som
              Mean        Std.Dev.    Min         Median      Max
  real        0.194       0.004       0.189       0.193       0.207
  user        0.259       0.035       0.200       0.260       0.310
  sys         0.015       0.013       0.000       0.010       0.050

The mean time has decreased by about 10% which is good, but the big question is
whether this is just due to our currently-very-slow GC or not. Using the JSON
benchmark as an example, even with GC fully turned off, there is a roughly 1%
improvement, so it does seem that, at least sometimes, it would be useful, even
with a "perfect" GC.