Enable the creation of variable sized SOM objects.
Previously we created objects on the stack and then moved them into the heap:
though dead-store elimination optimised this away, it meant that we were
constrained to fixed sized objects.

This commit expands the object creation API so that users can request enough
memory for an object *plus* extra memory which they can use for an
arbitrary-sized store. In essence, this gives us the flexibility
arbitrarily-sized structs without the difficulties that these require in normal
Rust.

As an example, we remove an allocation from SOM instances, which now store their
fields (of which there are a fixed number) immediately after the `Inst` object.
so that (on a 64-bit machine) the complete thing looks as follows:

  0:     vtable ptr: *const u8
  8:     class: Val
  16...: inst_vars

On the JSON benchmark this is about 2-3% faster with GC_DONT_GC=true and about
7-8% faster with the default GC options.