Make `SOMStack::split_off` return a `NormalArray` rather than a `Vec`.
This will allows us to remove an intermediate allocation and copy. Unfortunately
it also gets us into a bit of trouble with the borrow checker. Ideally we'd like
to write the following in the core VM:

  self.som_stack.split_off(self, ..)

but that borrows `self` mutably (`self.som_stack`) and immutably (the first
argument `self`). Fortunately `NormalArray::from_vec` doesn't need `VM` so we
can avoid this. Will this come back to bite us one day? We'll find out...