Split `Frame::new` into two.
This function is one of the most frequently called in yksom, so the fact that it
serves two different masters (methods and blocks) is less than ideal (at the
very least it introduces a branch). This commit splits the function into two,
one for methods and one for blocks.

To some extent this is a bit pointless in that the major overhead of
`Frame::new` is the malloc of a `Vec`, but one day that cost will be much less,
so this change seems like a useful one, if perhaps a bit early!