Lazily store the number of Unicode characters a string contains.
Calculating the number of Unicode characters in a string is an O(n) operation.
Previously everytime `String_::length` was called we iterated over the
underlying string to discover how many Unicode characters it contains.

This commit adds a field to SOM strings that stores its number of Unicode
characters. Since not all strings will be queried for this property, we lazily
cache this property the first time that `length` is called on a `String_`. We
use `usize::max` as our marker as that means the marker is "safe" in all cases.

On the JSON benchmark, with GC off, this speeds things up by about 80%.