Remove calls to tobj() when we know we must have a GCBOX.
At the moment a `Val` can only be `INT` or `GCBOX`, so in cases where we've
checked to see if `self` is an `INT`, the only remaining case is `GCBOX`. The
locations changed in this commit implicitly checked that `self` was a `GCBOX`
twice: by using unsafe `gcbox_to_tobj` function, we can remove the second
(pointless) check. This is a just about measurable improvement on the JSON
benchmark at about 1%.