The method a `Block` is contained within can never change.
`BlockInfo`s are (sort of) instantiated at run-time to make `Block`s. In order
to get proper backtraces, we have to track which method is executing, including
which method a block is part of. However, we previously stored the method on
`Block` even though it could never change at run-time.

This commit stores a block's method on `BlockInfo` so that we don't have to keep
passing it around at run-time. Because we instantiate `Block` so often, saving a
machine-word in this struct shaves ~2% off execution time of the JSON benchmark!