Separate out local variable lookup/set from up-variables.
"Up variables" is the Lua terminology for "variables in an outer closure". By
separating out variables local to the current closure from those in an outer
closure, we do get a small speed increase (3-4% on the JSON benchmark), but more
importantly we start to move the VM in a direction where we can put variables on
the stack.