Unify the internal representation of methods and blocks.
SOM methods and blocks are quite different from a user perspective. Previously
that internal representation leeched heavily into yksom's internals, causing an
awkward API, where we passed details such as the number of local variables
around explicitly.

This commit unifies the internal representation into a new type `Function` which
represents the common parts of methods and blocks internally to the VM. In other
words, methods and blocks now store `Function`s so that the core VM only has to
know about `Function`s and (not very much) about methods or blocks. The
abstraction isn't perfect: in particular, because methods can be SOM primitives,
not all `Function`s make use of all `Function`'s fields.

However, overall, and despite the monstrous size of this commit, this simplifies
yksom's internals and opens up future optimisation possibilities. Notably, we
can now track all of a function's blocks, which we were not able to do before.