Cache a function's blocks.
In functions with tight loops, continually allocating `Block`s (and `UpVar`s) is
expensive. It's also unnecessary. Consider this contrived function:

    f = ( | x |
        x := 5.
        [x = 0] whileFalse: [
            (x > 2) ifTrue: [x println].
            x := x - 1.
        ].
    )

Each time `ifTrue` executes the inner block, it captures the same variables from
the outer scope and operates on the same instance: there's no need to
continually create a fresh block. This PR thus caches blocks, storing them on
the SOM stack just after a function's variables.

How effective this is will depend hugely on a benchmark (a variant of the above
program can easily see 30-40% improvement), but it seems to be roughly a 5-10%
improvement on realistic benchmarks.