Remove the Vec allocation in `Block`.
This is our now "standard" optimisation. However, it can, and does create,
cycles which then causes Boehm to warn (endlessly...) that:

  GC Warning: Finalization cycle involving 0x7fa72b06aea0

In a sense this is really a problem for rboehm, because yksom doesn't define any
`Drop` methods, and isn't in danger of violating the "don't call (directly or
indirectly) anything with `Gc` in it in a `drop` method". However, until that's
fixed, this commit is unmergeable.