Null entries on the SOMStack after they're removed.
This prevents the GC thinking they're still alive. It would be preferable to
tell the GC "only scan the first N words of the stack" but we don't have any way
of doing that at the moment AFAIK.

This doesn't have any performance difference that I can measure either way.