Update for symmetric `=`.
This commit updates us to the latest SOM library and also adapts to the
new semantics of `=` for `String`/`Symbol` (see
https://github.com/SOM-st/SOM/pull/106).

Ironically, this sort-of already worked, except we previously made `=`
and `==` identical for strings/symbols. I believe that most (all?) SOMs
currently cache symbols so two instances of `#a` in different parts of
the source code always point to the same underlying object, so `==`
(nominally "reference equality") will return true when those two
seemingly separate objects are compared. yksom does not cache symbols,
so we have to override `==` to compare `Symbol`s by their contents. This
commit also does that for `String`s -- I don't think the SOM spec says
anything about this either way, and all the tests pass so...